chore: create omnichannel-typings package
