+++ packages/core-typings/src/IMessage/IMessage.ts
@@ -5,7 +5,6 @@ import type { Root } from '@rocket.chat/message-parser';
 import type { MessageSurfaceLayout } from '@rocket.chat/ui-kit';
 
 import type { ILivechatPriority } from '../ILivechatPriority';
-import type { ILivechatVisitor } from '../ILivechatVisitor';
 import type { IOmnichannelServiceLevelAgreements } from '../IOmnichannelServiceLevelAgreements';
 import type { IRocketChatRecord } from '../IRocketChatRecord';
 import type { IRoom, RoomID } from '../IRoom';
@@ -307,41 +306,6 @@ export interface IMessageReactionsNormalized extends IMessage {
 	};
 }
 
-export interface IOmnichannelSystemMessage extends IMessage {
-	navigation?: {
-		page: {
-			title: string;
-			location: {
-				href: string;
-			};
-			token?: string;
-		};
-	};
-	transferData?: {
-		comment: string;
-		transferredBy: {
-			name?: string;
-			username: string;
-		};
-		transferredTo: {
-			name?: string;
-			username: string;
-		};
-		nextDepartment?: {
-			_id: string;
-			name?: string;
-		};
-		scope: 'department' | 'agent' | 'queue';
-	};
-	requestData?: {
-		type: 'visitor' | 'user';
-		visitor?: ILivechatVisitor;
-		user?: Pick<IUser, '_id' | 'name' | 'username' | 'utcOffset'> | null;
-	};
-	webRtcCallEndTs?: Date;
-	comment?: string;
-}
-
 export type IVoipMessage = IMessage & {
 	voipData: {
 		callDuration?: number;
@@ -410,9 +374,3 @@ export type IMessageWithPendingFileImport = IMessage & {
 		downloaded?: boolean;
 	};
 };
-
-export interface IMessageFromVisitor extends IMessage {
-	token: string;
-}
-
-export const isMessageFromVisitor = (message: IMessage): message is IMessageFromVisitor => 'token' in message;
