--- packages/core-typings/src/omnichannel/index.ts
@@ -1,5 +1,0 @@
-export * from './sms';
-export * from './routing';
-export * from './queue';
-export * from './reports';
-export * from './mac';
