+++ packages/omnichannel-typings/package.json
@@ -1,0 +1,31 @@
+{
+	"$schema": "https://json.schemastore.org/package",
+	"name": "@rocket.chat/omnichannel-typings",
+	"version": "6.14.0-develop",
+	"devDependencies": {
+		"@rocket.chat/apps-engine": "workspace:^",
+		"@rocket.chat/eslint-config": "workspace:^",
+		"eslint": "~8.45.0",
+		"mongodb": "^4.17.2",
+		"prettier": "~2.8.8",
+		"typescript": "~5.5.4"
+	},
+	"scripts": {
+		"lint": "eslint --ext .js,.jsx,.ts,.tsx .",
+		"lint:fix": "eslint --ext .js,.jsx,.ts,.tsx . --fix",
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"dev": "tsc --watch --preserveWatchOutput -p tsconfig.json",
+		"build": "rm -rf dist && tsc -p tsconfig.json"
+	},
+	"main": "./dist/index.js",
+	"typings": "./dist/index.d.ts",
+	"files": [
+		"/dist"
+	],
+	"dependencies": {
+		"@rocket.chat/core-typings": "workspace:^"
+	},
+	"volta": {
+		"extends": "../../package.json"
+	}
+}
