+++ packages/omnichannel-typings/src/IInquiry.ts
@@ -1,9 +1,9 @@
-import type { ILivechatPriority } from '../ILivechatPriority';
+import type { IMessage, IRocketChatRecord } from '@rocket.chat/core-typings';
+
+import type { ILivechatPriority } from './ILivechatPriority';
 import type { ILivechatVisitor } from './ILivechatVisitor';
-import type { IMessage } from '../IMessage';
-import type { IOmnichannelServiceLevelAgreements } from '../IOmnichannelServiceLevelAgreements';
-import type { IRocketChatRecord } from '../IRocketChatRecord';
-import type { IOmnichannelRoom, OmnichannelSourceType } from '../IRoom';
+import type { IOmnichannelServiceLevelAgreements } from './IOmnichannelServiceLevelAgreements';
+import type { IOmnichannelRoom, OmnichannelSourceType } from './Room';
 import type { SelectedAgent } from './routing';
 
 export interface IInquiry {
