+++ packages/omnichannel-typings/src/ILivechatAgentActivity.ts
