+++ packages/omnichannel-typings/src/ILivechatBusinessHour.ts
