+++ packages/omnichannel-typings/src/ILivechatContact.ts
