+++ packages/omnichannel-typings/src/ILivechatCustomField.ts
