+++ packages/omnichannel-typings/src/ILivechatDepartment.ts
