+++ packages/omnichannel-typings/src/ILivechatDepartmentAgents.ts
