+++ packages/omnichannel-typings/src/ILivechatDepartmentRecord.ts
