+++ packages/omnichannel-typings/src/ILivechatTrigger.ts
