+++ packages/omnichannel-typings/src/IOmnichannel.ts
