+++ packages/omnichannel-typings/src/IOmnichannelCannedResponse.ts
