+++ packages/omnichannel-typings/src/mac.ts
